Finish monthly report view + subviews split
