+++ Android/EasyBudget/app/src/main/java/com/benoitletondor/easybudgetapp/view/monthlyreport/subviews/EmptyView.kt
@@ -1,0 +1,47 @@
+package com.benoitletondor.easybudgetapp.view.monthlyreport.subviews
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.benoitletondor.easybudgetapp.R
+
+@Composable
+fun EmptyView() {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(horizontal = 20.dp, vertical = 20.dp),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        Image(
+            painter = painterResource(R.drawable.ic_date_grey_48dp),
+            contentDescription = null,
+        )
+
+        Spacer(modifier = Modifier.height(15.dp))
+
+        Text(
+            modifier = Modifier.fillMaxWidth(),
+            text = stringResource(R.string.monthly_report_no_entries_placeholder),
+            color = colorResource(R.color.placeholder_text),
+            fontSize = 14.sp,
+            textAlign = TextAlign.Center,
+        )
+    }
+}
